Move product form bottom sheet actions calculation to a separate struct `ProductFormBottomSheetActionsFactory`. Move existing unit tests to the factory.
