+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/ProductFormViewController.swift
@@ -229,11 +229,6 @@ private extension ProductFormViewController {
     }
 
     func configureMoreDetailsContainerView() {
-        guard isEditProductsRelease2Enabled else {
-            moreDetailsContainerView.isHidden = true
-            return
-        }
-
         let title = NSLocalizedString("Add more details", comment: "Title of the button at the bottom of the product form to add more product details.")
         let viewModel = BottomButtonContainerView.ViewModel(style: .link,
                                                             title: title,
@@ -258,8 +253,10 @@ private extension ProductFormViewController {
         let title = NSLocalizedString("Add more details",
                                       comment: "Title of the bottom sheet from the product form to add more product details.")
         let viewProperties = BottomSheetListSelectorViewProperties(title: title)
-        let dataSource = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                   isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { [weak self] action in
+        let actions = ProductFormBottomSheetActionsFactory.actions(product: product,
+                                                                   isEditProductsRelease2Enabled: isEditProductsRelease2Enabled,
+                                                                   isEditProductsRelease3Enabled: isEditProductsRelease3Enabled)
+        let dataSource = ProductFormBottomSheetListSelectorCommand(actions: actions) { [weak self] action in
                                                                     self?.dismiss(animated: true) { [weak self] in
                                                                         switch action {
                                                                         case .editInventorySettings:
@@ -282,16 +279,10 @@ private extension ProductFormViewController {
     }
 
     func updateMoreDetailsButtonVisibility(product: Product) {
-        guard isEditProductsRelease2Enabled else {
-            moreDetailsContainerView.isHidden = true
-            return
-        }
-
-        let moreDetailsActions: [ProductFormBottomSheetAction] = isEditProductsRelease3Enabled ?
-            [.editInventorySettings, .editShippingSettings, .editCategories, .editBriefDescription]:
-            [.editInventorySettings, .editShippingSettings, .editBriefDescription]
-        let hasVisibleActions = moreDetailsActions.map({ $0.isVisible(product: product) }).contains(true)
-        moreDetailsContainerView.isHidden = hasVisibleActions == false
+        let moreDetailsActions = ProductFormBottomSheetActionsFactory.actions(product: product,
+                                                                              isEditProductsRelease2Enabled: isEditProductsRelease2Enabled,
+                                                                              isEditProductsRelease3Enabled: isEditProductsRelease3Enabled)
+        moreDetailsContainerView.isHidden = moreDetailsActions.isEmpty
     }
 }
