+++ WooCommerce/WooCommerceTests/ViewRelated/Products/BottomSheetListSelector/ProductFormBottomSheetListSelectorCommandTests.swift
@@ -3,103 +3,13 @@ import XCTest
 @testable import Yosemite
 
 final class ProductFormBottomSheetListSelectorCommandTests: XCTestCase {
-    // MARK: - `data`
-
-    // M3 feature flag off
-
-    func testDataHasNoEditProductsRelease3ActionsForAPhysicalProductWhenFeatureFlagIsOff() {
-        let product = Fixtures.physicalProduct
-        let isEditProductsRelease3Enabled = false
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editShippingSettings,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
-    func testDataHasNoEditProductsRelease3AndShippingActionsForAVirtualProductWhenFeatureFlagIsOff() {
-        let product = Fixtures.virtualProduct
-        let isEditProductsRelease3Enabled = false
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
-    func testDataHasNoEditProductsRelease3AndShippingActionsForADownloadableProductWhenFeatureFlagIsOff() {
-        let product = Fixtures.downloadableProduct
-        let isEditProductsRelease3Enabled = false
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
-    // M3 feature flag on
-
-    func testDataHasEditProductsRelease3ActionsForAPhysicalProductWhenFeatureFlagIsOn() {
-        let product = Fixtures.physicalProduct
-        let isEditProductsRelease3Enabled = true
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editShippingSettings,
-            .editCategories,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
-    func testDataHasEditProductsRelease3ButNoShippingActionsForAVirtualProductWhenFeatureFlagIsOn() {
-        let product = Fixtures.virtualProduct
-        let isEditProductsRelease3Enabled = true
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editCategories,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
-    func testDataHasEditProductsRelease3ButNoShippingActionsForADownloadableProductWhenFeatureFlagIsOn() {
-        let product = Fixtures.downloadableProduct
-        let isEditProductsRelease3Enabled = true
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: isEditProductsRelease3Enabled) { _ in }
-
-        let expectedActions: [ProductFormBottomSheetAction] = [
-            .editInventorySettings,
-            .editCategories,
-            .editBriefDescription
-        ]
-        XCTAssertEqual(command.data, expectedActions)
-    }
-
     // MARK: - `handleSelectedChange`
 
     func testCallbackIsCalledOnSelection() {
         // Arrange
-        let product = Fixtures.physicalProduct
+        let actions: [ProductFormBottomSheetAction] = [.editInventorySettings, .editShippingSettings, .editCategories, .editBriefDescription]
         var selectedActions = [ProductFormBottomSheetAction]()
-        let command = ProductFormBottomSheetListSelectorCommand(product: product,
-                                                                isEditProductsRelease3Enabled: true) { selected in
+        let command = ProductFormBottomSheetListSelectorCommand(actions: actions) { selected in
                                                                     selectedActions.append(selected)
         }
 
@@ -117,25 +27,5 @@ final class ProductFormBottomSheetListSelectorCommandTests: XCTestCase {
             .editCategories
         ]
         XCTAssertEqual(selectedActions, expectedActions)
-    }
-}
-
-private extension ProductFormBottomSheetListSelectorCommandTests {
-    enum Fixtures {
-        // downloadable: false, virtual: false, missing inventory/shipping/categories/brief description
-        static let physicalProduct = MockProduct().product(downloadable: false, briefDescription: "", manageStock: true, sku: nil, stockQuantity: nil,
-                                                           dimensions: ProductDimensions(length: "", width: "", height: ""), weight: nil,
-                                                           virtual: false,
-                                                           categories: [])
-        // downloadable: false, virtual: true, missing inventory/shipping/categories/brief description
-        static let virtualProduct = MockProduct().product(downloadable: false, briefDescription: "", manageStock: true, sku: nil, stockQuantity: nil,
-                                                          dimensions: ProductDimensions(length: "", width: "", height: ""), weight: nil,
-                                                          virtual: true,
-                                                          categories: [])
-        // downloadable: true, virtual: true, missing inventory/shipping/categories/brief description
-        static let downloadableProduct = MockProduct().product(downloadable: true, briefDescription: "", manageStock: true, sku: nil, stockQuantity: nil,
-                                                               dimensions: ProductDimensions(length: "", width: "", height: ""), weight: nil,
-                                                               virtual: true,
-                                                               categories: [])
     }
 }
